Top nav , left nav, and player grabed into a card model
