+++ components/CategoryPage/index.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const CategoryPage = () => {
+  return <div>CategoryPage</div>;
+};
+
+export default CategoryPage;
