+++ components/LatestPage/index.jsx
@@ -1,34 +1,23 @@
-import React from 'react'
-import BottomPlayer from '../BottomPaler/BottomPalyer'
-import HotSongs from '../modules/HotSongs'
-import { LeftNav, TopNav } from '../modules/NavBar'
+import React from "react";
+import HotSongs from "../modules/HotSongs";
 import { leftNavPinedAtom } from "../lib/atoms";
-import { useRecoilValue } from 'recoil';
-import TopSongs from '../modules/TopSongs/TopSongs';
-
+import { useRecoilValue } from "recoil";
+import TopSongs from "../modules/TopSongs/TopSongs";
+import PageCard from "../modules/__modules__/Card/PageCard";
 
 const LatestPage = () => {
   const isLeftNavPined = useRecoilValue(leftNavPinedAtom);
 
   return (
     <div className="bg-globalBg">
-    <div
-      className={`relative flex flex-col ${
-        !isLeftNavPined ? "ml-28 transition-all" : "transition-all ml-72"
-      } mobile:ml-5`}
-    >
-      <TopNav />
-      <div className='mt-20'>
-      <TopSongs/>
-      
-      <HotSongs />
-      </div>
+      <PageCard>
+        <div className="mt-20">
+          <TopSongs />
+          <HotSongs />
+        </div>
+      </PageCard>
     </div>
-    <LeftNav />
-    <BottomPlayer />
-  </div>
-  )
+  );
+};
 
-}
-
-export default LatestPage
+export default LatestPage;
