+++ components/modules/Category/index.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Category = () => {
+  return <div>Category</div>;
+};
+
+export default Category;
